requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.0
face-recognition==1.3.0
numpy==1.24.3
scikit-learn==1.3.2
pillow==10.1.0
werkzeug==2.3.7
python-dotenv==1.0.0
setuptools<81.0.0